ActorService soap implementation fix + MovieService getMovie now retrieves cast data

+++ services/movieservice/src/main/java/it/uniflix/movieservice/controller/rest/GenreRestApi.java
@@ -33,7 +33,7 @@ public interface GenreRestApi {
 			})
 	@GET
 	@Path("{id}/movies")
-	public List<Movie> getMoviesByGenre(@PathParam(value = "id") int id);
+	public List<Movie> getMoviesByGenre(@PathParam(value = "id") long id);
 	
 	@Operation(
 			description = "Retrieve all genres",
